Added a potentiometer and a loop button.
To implement the effect switch, the "Potentiometer" class
was added, which is an inheritor of the "QDial" class.
The potentiometer has 4 positions to select one of the
four subsequent effects.

Also added was the "LoopButton" class, which is an inheritor of
"QPushButton", to implement a button for looping the track.