Adding new classes for setting up audio track
The following classes have been added:
    1) TrackColorButtons - a class implementing a button for selecting a color
for a track button. It has a similar appearance to a track, including
2 zones: the central (a circle showing the active state track) and the
outer (just for visual recognition). The lower zone can be given a color.
The corresponding changes send a signal that is processed in TrackSettings
and then appears in the track.
    2) Potentiometer - a class implementing an effects switch
(designed for 4 positions). It is made in the style of a round potentiometer,
like on regular mixing consoles. Has 4 rating zones:
        0 (10:31 - 1:30) - setFirstQuarter
        25 (1:31 - 4:30) - setSecondQuarter
        50 (4:31 - 7:30) - setThird Quarter
        75 (7:30 - 10:30) - setFourth Quarter
    3) ClickableLabel - a girl who implements an inscription that can be
clicked on, and it generates a signal. ClickableLabel is used to indicate
the named effects under the potentiometer. In the Track Settings window,
the potentiometer is associated with 4 "ClickableLabel" inscriptions.
When setting the potentiometer in a zone, the corresponding inscription
(effect name) lights up. When clicking on the inscription (effect name),
the potentiometer must be in the position.