Adding new classes for audio track settings
The following classes have been added:
    1) AudioSampleSelector - a class implementing a button for selecting
an audio sample for a track. It has a look that is defined by two
cards that display 2 states: selected/not selected with a track.
By default, this element has the look of a 6.3 mm Jack connector.
When selecting an audio sample, a second picture is set, which has
the look of a 6.3 mm Jack connector in some connected wire.
    2) LoopButton - a class implementing a track loop switch.
It looks like a simple gray square button, when pressed,
it changes its color to red.

The corresponding signals about changing the frequency of elements are associated with TrackSettings, which is then associated with the track.