Changes to the window settings and its connection with the track
The settings window has been thoroughly redesigned. All widgets for
manipulating track settings have been replaced with corresponding user widgets.

But a quick replacement led to a significant drop in performance when
enabled with track settings window widgets. Presumably, this happened
due to the large number of signals.

The fact is that previously each track had its own window settings and
was directly connected to them: in the track designer, connections were
established with window settings signals.

But replacing standard widgets in the settings window with a user showed that
this approach is ineffective and greatly reduces performance. This started
even when setting up a 3 by 3 project (9 tracks and in each settings window)

The equipment on which the technology was developed for myself:
    Processor: Intel Core i3 9100F
    RAM: 16 GB DDR4 2666 MHz
    GPU: AMD Radeon R7 250 1 GB DDR5

Therefore, the settings for linking the track and the window were changed. Now there
is one instance of the window air conditioner, which is a problem of the
"Project" class. In the "Project" class, a void openTrackSettings(Track* track)
method was added, which takes a pointer to a track object and links it to the window
settings, redefining the parameters of the received track in the window and linking the
received track to the window signals. Before transferring the track parameters to the
window and linking the track to the window, the old connection windows are previously deleted.
Accordingly, all the logic for linking a track to the settings window was removed from
the Track class, and corresponding getters/setters were added to get and set track parameters.

These changes have significantly increased performance.
Additionally, minor changes were made to correct warnings.