Bug fixes
1) Now if a track is looped and active and the audio sample selector
is also active, then when you click on the audio sample selector
(AudioSampleSelector), it is reset as before (becomes inactive) and
the track stops (stops playing the audio sample)

2) Some fields were removed from the Project class and transferred to
the constructor as local pointers to dynamic memory owned by the central
widgets, which are fields of the class (Project). A field for the project
name was also added.

3) In the Track class, in addition to the improved connection with
(AudioSampleSelector) (described above, in point 1), changes were made
to the basic colors of the buttons in the constructor (now the central
indicator of the button has a red active color and a gray passive color
by default)

4) In the (TrackSettings) class, the fields responsible for storing the
set sixteenth parts of the bar and the corresponding signals were removed.
The initial values ??for the fader displays set in the constructor were
also corrected.