Recording implementation start and some improvements
1) Added a flag (m_is_recording) to the Track class, which corresponds
to the state of the note.

2) Fixed a track bug: if the track was looped, then after pressing the
loop button to change this mode, the track still remained partially
looped - it stopped playing sound at the beginning of a new measure.
    Now, when switching a track from looped to unlooped state, it
continues to behave only as looped until the next new measure. After
a new measure, the track is completely unlooped.

3) Added a button to the TrackSettings class to initialize recording
of the selected track.

4) The logic of the record buttons and the record timer has been moved from the REC class to the Project class